Add instruction explanation. Add demo image

+++ source/display.c
@@ -44,7 +44,7 @@ void mv_set_line_width(mv_display *display, uint16_t width)
  * @param x1 The x coordinate of the end point
  * @param y1 The y coordinate of the end point
  */
-void mv_draw_line(mv_display *display, int32_t x0, int32_t y0, int32_t x1, int32_t y1)
+void mv_draw_line(mv_display *display, int16_t x0, int16_t y0, int16_t x1, int16_t y1)
 {
     if (display->point_count >= MV_MAX_DISPLAY_POINTS)
     {
